+++ .github/workflows/ci.yml
@@ -4,6 +4,7 @@ env:
   AWS_EC2_METADATA_DISABLED: true
   S3_RELEASE_PREFIX: "s3://redgold-public/release"
   DOCKER_USER: "redgoldio"
+  TOOLCHAIN: "nightly"
 on: push
 jobs:
   test:
@@ -28,7 +29,7 @@ jobs:
       - uses: actions-rs/toolchain@v1
         with:
           profile: minimal
-          toolchain: stable
+          toolchain: ${{ env.TOOLCHAIN }}
       - uses: Swatinem/rust-cache@v1
         with:
           cache-on-failure: true
@@ -62,7 +63,7 @@ jobs:
       - uses: actions-rs/toolchain@v1
         with:
           profile: minimal
-          toolchain: stable
+          toolchain: ${{ env.TOOLCHAIN }}
       - uses: Swatinem/rust-cache@v1
         with:
           cache-on-failure: true
